+++ src/main/java/online/kingdomkeys/kingdomkeys/network/PacketHandler.java
@@ -57,6 +57,7 @@ public class PacketHandler {
 		client(SCSyncOrganizationData.TYPE, SCSyncOrganizationData.STREAM_CODEC);
 		client(SCSyncOrgPortalPacket.TYPE, SCSyncOrgPortalPacket.STREAM_CODEC);
 		client(SCSyncOrbStats.TYPE, SCSyncOrbStats.STREAM_CODEC);
+		client(SCStartReversal.TYPE, SCStartReversal.STREAM_CODEC);
 		client(SCSyncPlayerData.TYPE, SCSyncPlayerData.STREAM_CODEC);
 		client(SCSyncShopData.TYPE, SCSyncShopData.STREAM_CODEC);
         client(SCSyncSellData.TYPE, SCSyncSellData.STREAM_CODEC);